Trying out better way to handle multiple MESSAGE_CREATE events prevent duplicate code and multiple of the same calls on each message

+++ app/Discord/Moderation/Events/MediaFilter.php
@@ -2,49 +2,26 @@
 
 namespace App\Discord\Moderation\Events;
 
-use App\Discord\Core\DiscordEvent;
-use Discord\Discord;
+use App\Discord\Core\Bot;
+use App\Discord\Core\Guild;
+use App\Discord\Core\MessageCreateEvent;
+use App\Models\Channel;
 use Discord\Parts\Channel\Message;
-use Discord\WebSockets\Event;
 
-class MediaFilter extends DiscordEvent
+class MediaFilter implements MessageCreateEvent
 {
-    /**
-     * @return void
-     */
-    public function registerEvent(): void
+    public function execute(Bot $bot, Guild $guild, Message $message, ?Channel $channel): void
     {
-        $this->discord->on(Event::MESSAGE_CREATE, function (Message $message, Discord $discord) {
-            if ($message->author->bot) {
+        if (!$channel || !$channel->media_only || $message->attachments->count() > 0) {
+            return;
+        }
+        // If the message contains a valid URL we allow it.
+        foreach (explode(' ', $message->content) as $word) {
+            if (filter_var($word, FILTER_VALIDATE_URL)) {
                 return;
             }
-            if (!$message->guild_id) {
-                return;
-            }
-            $guild = $this->bot->getGuild($message->guild_id);
-            if ($guild) {
-                $channel = $guild->getChannel($message->channel_id);
-                if (!$channel) {
-                    return;
-                }
-                if (!$channel->media_only) {
-                    return;
-                }
-                // If message contains Images, audio or any other file we allow it
-                if ($message->attachments->count() > 0) {
-                    return;
-                }
-                // If the message contains a valid URL we allow it.
-                foreach (explode(' ', $message->content) as $word) {
-                    if (filter_var($word, FILTER_VALIDATE_URL)) {
-                        return;
-                    }
-                }
-                $message->delete();
-                $message->author->sendMessage(__('bot.media-deleted', ['channel' => $message->channel->name]));
-            }
-        });
+        }
+        $message->delete();
+        $message->author->sendMessage(__('bot.media-deleted', ['channel' => $message->channel->name]));
     }
-
-
 }
